Share the address-association rules instead of importing the watcher
`backfill_address_events()` reached into `apiwatcher` through a local
import, which closed a direct `apiwatcher -> dbbuilder -> apiwatcher`
cycle that pylint flagged. `EVENTS_ADDRESS_FIELDS` and the two functions
that write `address_events` now live in `api/addressevents.py`, imported by
the three places that have to agree on those rules -- the watcher, migration
0001 and the backfill -- and by `parser/mempool.py`, which previously pulled
in the whole watcher module for one dict.

Also silence two pylint alerts on the build hook: hatchling is a build-time
dependency and is absent from the lint environment, and `version` is part of
the build-hook interface.

The remaining `apiwatcher -> dbbuilder -> staterollback` cycle predates this
branch (pylint reports it on develop too, anchored at `wsgi.py`); breaking it
means moving the watcher's parsed-event query layer, which does not belong in
this PR.
